Python: Move more tests for argument names into own file
Plus fixup of expected output from unrelated tests
